+++ erpnext_egypt_compliance/erpnext_eta/doctype/eta_connector/eta_connector.js
