+++ erpnext_egypt_compliance/erpnext_eta/doctype/eta_settings/eta_settings.js
