+++ erpnext_egypt_compliance/erpnext_eta/doctype/eta_tax_type/eta_tax_type.js
