+++ erpnext_egypt_compliance/erpnext_eta/report/eta_sales_invoices_status/eta_sales_invoices_status.py
@@ -2,7 +2,7 @@
 # For license information, pleASe see license.txt
 
 import frappe
-from erpnext_egypt_compliance.erpnext_egypt_compliance.utils import get_company_eta_connector
+from erpnext_egypt_compliance.erpnext_eta.utils import get_company_eta_connector
 from datetime import datetime
